feat: give each tour season its own /tour/[year] route
